Dev Commit
[ Fixes ]
- Fixed where Resetting the settings Crashes the app.
- Fixed API timeout was accidentally set to 2 seconds instead of 10 seconds.
- Fixed Location textbox focus outline when tabbed.

[ New | Changes ]
- Added Custom Background option.
- Added SelectionPopup's Closing Animation when game is clicked.

+++ src/HoyoLauncher.Core/Config/MainConfig.cs
@@ -17,10 +17,10 @@ public sealed class MainConfig
     public string ZZZ_DIR { get; set; }
 
     /// <summary> Last game after close </summary>
-    [DefaultValue(0)]
+    [DefaultValue((short)0)]
     public short LAST_GAME { get; set; }
 
-    [DefaultValue(1)]
+    [DefaultValue((short)1)]
     public short EXIT_MODE { get; set; }
     [DefaultValue(true)]
     public bool CHECKBOX_BACKGROUND { get; set; }
